+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Add SingleLineThrowFixer
 - Add PhpUnitNoUselessReturnFixer
 - Feature: NoCommentedOutCodeFixer - handle class method
+- Deprecate SingleLineThrowFixer
 
 ## v1.13.0 - *2019-01-11*
 - Update PHP CS Fixer to v2.14